Move _tau, improve coverage
