+++ src/geom/type2.jl
@@ -39,15 +39,30 @@ for a 2D parallel or fan-beam system.
 * `d`, `source_offset`, `dsd`, `dod`
   must all have the same units.
 
-# Derived values (non-exported helper functions):
+# Basic methods
+
+* `angles` (na) in degrees
+* `dims (nb, na)`
+* `ones = ones(Float32, nb, na)`
+* `zeros = zeros(Float32, nb, na)`
+* `rays` iterator of `(r, ϕ)` parallel-beam coordinate tuples of size `(nb, na)`
+* `downsample(st, down)` reduce sampling by integer factor
+* `oversample(st, over)`
+* `sino_geom_plot!` plot system geometry
+
+# Non-exported helper functions for developers:
 
 * `_ds|dr` radial sample spacing (`NaN` for `:moj`)
 * `_s (nb) s` sample locations
 * `_w = (nb-1)/2 + offset` "middle" sample position
-* `_ar (na)` source angles `radians`
+* `_ar (na)` source angles [radians]
 * `_rfov` radial FOV
 * `_xds (nb)` center of detector elements (beta=0)
 * `_yds (nb)` ""
+* `_tau(rg, x, y)` projected s/ds for each `(x,y)` pair `(length(x), na)`
+* `_shape(rg, sino [,:])` reshape `sino` into array `(nb,na[,:])`
+* `_unitv(rg [, (ib,ia)])`
+  unit 'vector' with single nonzero element
 
 For mojette:
 
@@ -59,25 +74,7 @@ For fan beam:
 * `_dfs` distance from source to detector focal spot
         (0 for 3rd gen CT, `Inf` for flat detectors)
 * `_gamma(rg [,s]) (nb)` gamma sample values `radians`, optionally given `s` values
-* `.gamma_max` half of fan angle `radians`, if `offset` == 0
-
-# Basic methods
-
-* `angles` (na) in degrees
-* `dims (nb, na)`
-* `ones = ones(Float32, nb, na)`
-* `zeros = zeros(Float32, nb, na)`
-* `rays` iterator of `(r, ϕ)` parallel-beam coordinate tuples of size `(nb, na)`
-* `downsample(st, down)` reduce sampling by integer factor
-* `oversample(st, over)`
-* `geom_plot!` plot system geometry
-
-# Methods
-
-* `_shape(rg,sino)` reshape sinograms into array `(nb,na,:)`
-* `_unitv(rg [, (ib,ia)])`
-  unit 'vector' with single nonzero element
-* `_taufun(rg,x,y)` projected s/ds for each (x,y) pair `(numel(x),na)`
+* `_gamma_max = max(|γ|)` half of fan angle `radians`, if `offset_s` == 0
 
 # Notes
 * Use `ct_geom()` instead for 3D axial or helical cone-beam CT.
